refactor: move wrap handler logic to mixin class
